+++ example/android/app/src/main/java/com/aotasoft/photo_view_example/MainActivity.java
@@ -1,6 +1,7 @@
 package com.aotasoft.photo_view_example;
 
 import io.flutter.embedding.android.FlutterActivity;
+import io.flutter.embedding.android.FlutterFragmentActivity;
 
-public class MainActivity extends FlutterActivity {
+public class MainActivity extends FlutterFragmentActivity {
 }