+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 
 ## 1.0.1
 * PhotoView aims to help produce an easily usable implementation of a zooming ImageView. Update import file
+
+## 1.0.2
+* PhotoView aims to help produce an easily usable implementation of a zooming ImageView. Update import file
+* Update android support swipe up/down to dismiss
+* Update indicator for android
+* Update initial start page